requirements.txt
datetime
newsapi-python